wip: criando tabela de users

+++ db/migrate/20250627031804_create_a_bairros.rb
@@ -2,17 +2,18 @@
 
 class CreateABairros < ActiveRecord::Migration[7.2]
   def up
-    create_table :a_bairros do |t|
-      t.string :descricao
-          
-      t.string :created_by
-      t.string :updated_by
-      t.datetime :deleted_at
-      t.timestamps
+    unless table_exists?(:a_bairros)
+      create_table :a_bairros do |t|
+        t.string   :descricao
+        t.string   :created_by
+        t.string   :updated_by
+        t.datetime :deleted_at
+        t.timestamps
+      end
     end
   end
 
   def down
-    drop_table :a_bairros
+    drop_table :a_bairros if table_exists?(:a_bairros)
   end
 end
